add speding by category graph.
