+++ src/state/slice/transactions/transactions.test.ts
@@ -13,5 +13,4 @@ describe("Transactions slice", () => {
 
     expect(result).toPartiallyContain(action.payload);
   });
-  it.todo("should remove Transaction");
 });
