+++ src/views/Info/SpendingByCategoryLineGraph/index.tsx
@@ -1,0 +1,1 @@
+export * from "./SpendingByCategoryLineGraph";
