--- src/views/Info/SpendingByWeekdayGraph/SpendingByWeekdayGraph.tsx
@@ -1,8 +1,0 @@
-import { useSelector } from "react-redux";
-import { transactionsSelector } from "state/selectors/transactionsSelector";
-
-export const SpendingByWeekdayGraph = () => {
-  const transactions = useSelector(transactionsSelector);
-
-  return <div>Test</div>;
-};
