--- src/views/Info/SpendingByWeekdayGraph/index.tsx
@@ -1,1 +1,0 @@
-export * from "./SpendingByWeekdayGraph";
